refactor(scout-for-lol): split durable rollback helpers
Why
The Temporal migration added durable rollback and lake fold logic to existing modules that exceeded the repository's max-lines quality gate.

What
- move Explore reservation rollback into its own module
- move report-lake fold parquet writing into its own module
- preserve existing imports through explicit caller updates

Verification
- cd packages/scout-for-lol/packages/backend && bun run typecheck
- cd packages/scout-for-lol/packages/backend && bun run lint (0 errors; existing duplication warnings remain)
